requirements.txt
Flask
python-dotenv
openai
deepgram-sdk
Werkzeug
pypdf2
python-docx 